# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "React Online Store",
        "version": "0.1"
    },
    "paths": {
...
        "/api/admin/products": {
            "get": {
                "tags": [
                    "AdminProduct"
                ],
                "summary": "Display a listing of the Products.",
                "operationId": "9058305dfe0377e6cf7c5253ed664401",
                "responses": {
                    "200": {
                        "description": "Display a listing of the Products",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProductResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "AdminProduct"
                ],
                "summary": "Store a newly created Product.",
                "operationId": "e78a4cd5a8e47d96daa7a5a7a13d3e30",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Stored a newly created Product",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/products/{id}": {
            "get": {
                "tags": [
                    "AdminProduct"
                ],
                "summary": "Display the specified Product by ID.",
                "operationId": "e645351e64571486935b1fbb5743d6b8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Display the specified Product by ID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "AdminProduct"
                ],
                "summary": "Update the specified Product by ID.",
                "operationId": "78c34d6552e97335a6594ceaf054d60f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated the specified Product by ID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "AdminProduct"
                ],
                "summary": "Remove the specified Product by ID.",
                "operationId": "14131d6e77dc1537765eaa35848d3941",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Removed the specified Product by ID"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/products/restore": {
            "post": {
                "tags": [
                    "AdminProduct"
                ],
                "summary": "Restore the specified Products by ID.",
                "operationId": "4d671f0d8b20222a0775f07ffd84fd84",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RestoreProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Restore products",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProductResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/products/{id}/force": {
            "delete": {
                "tags": [
                    "AdminProduct"
                ],
                "summary": "Remove the specified Product by ID.",
                "operationId": "0e38ce21db896c0371ed2b703d5348f1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Removed the specified Product by ID"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Display the specified Product by ID.",
                "operationId": "c87fb2e4d617e7b36c712cad358acdb8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Display the specified Product by ID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "RestoreProductRequest": {
                "required": [
                    "id"
                ],
                "properties": {
                    "id": {
                        "description": "Products list for restore",
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    }
                },
                "type": "object"
            },
            "StoreProductRequest": {
                "required": [
                    "name",
                    "price"
                ],
                "properties": {
                    "name": {
                        "description": "Product name",
                        "type": "string",
                        "maxLength": 255
                    },
                    "price": {
                        "description": "Product price",
                        "type": "number",
                        "maximum": 10000,
                        "minimum": 0
                    },
                    "description": {
                        "description": "Product description",
                        "type": "string",
                        "maxLength": 10000
                    },
                    "available": {
                        "description": "Product available qty",
                        "type": "integer",
                        "maximum": 1000,
                        "minimum": 0
                    },
                    "sale": {
                        "description": "Product sale %",
                        "type": "number",
                        "maximum": 100,
                        "minimum": 0
                    }
                },
                "type": "object"
            },
            "UpdateProductRequest": {
                "properties": {
                    "name": {
                        "description": "Product name",
                        "type": "string",
                        "maxLength": 255
                    },
                    "price": {
                        "description": "Product price",
                        "type": "number",
                        "maximum": 10000,
                        "minimum": 0
                    },
                    "description": {
                        "description": "Product description",
                        "type": "string",
                        "maxLength": 10000
                    },
                    "available": {
                        "description": "Product available qty",
                        "type": "integer",
                        "maximum": 1000,
                        "minimum": 0
                    },
                    "sale": {
                        "description": "Product sale %",
                        "type": "number",
                        "maximum": 100,
                        "minimum": 0
                    }
                },
                "type": "object"
            },
...
            "BrandResource": {
                "properties": {
                    "id": {
                        "description": "Class BrandResource",
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "image": {
                        "$ref": "#/components/schemas/ImageResource"
                    }
                },
                "type": "object"
            },
            "CategoryResource": {
                "properties": {
                    "id": {
                        "description": "Class CategoryResource",
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "image": {
                        "$ref": "#/components/schemas/ImageResource"
                    }
                },
                "type": "object"
            },
            "ImageResource": {
                "properties": {
                    "id": {
                        "description": "Class ImageResource",
                        "type": "integer"
                    },
                    "path": {
                        "type": "string"
                    },
                    "origin_name": {
                        "type": "string"
                    },
                    "width": {
                        "type": "string"
                    },
                    "height": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ProductResource": {
                "properties": {
                    "id": {
                        "description": "Class ProductResource",
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "sku": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number"
                    },
                    "description": {
                        "type": "string"
                    },
                    "available": {
                        "type": "integer"
                    },
                    "sale": {
                        "type": "number"
                    },
                    "category": {
                        "$ref": "#/components/schemas/CategoryResource"
                    },
                    "brand": {
                        "$ref": "#/components/schemas/BrandResource"
                    },
                    "image": {
                        "$ref": "#/components/schemas/ImageResource"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "bearerFormat": "jwt",
                "scheme": "bearer"
            }
        }
    }
}